frontend/uploadapi/requirements.txt
aniso8601==1.3.0
backports.shutil-get-terminal-size==1.0.0
certifi==2017.11.5
chardet==3.0.4
click==4.1
decorator==4.1.2
enum34==1.1.6
Flask==0.12.2
Flask-Cors==3.0.3
Flask-RESTful==0.3.6
flask-restplus==0.10.1
flask-shell-ipython==0.3.0
functools32==3.2.3.post2
hashids==1.2.0
httpie==0.9.9
idna==2.6
ipdb==0.10.3
ipython==5.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
logging-tree==1.7
MarkupSafe==1.0
marshmallow==2.14.0
neo4j-driver==1.1.0rc1
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
py2neo==4.0.0b2
Pygments==2.2.0
python-dateutil==2.6.1
python-magic==0.4.15
pytz==2017.2
requests==2.18.4
scandir==1.6
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.12.2
